requirements.txt
pygame
pysimplegui
quick-mailer